requirements.txt
astroid==2.3.2
autopep8==1.4.4
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
future==0.18.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.3
python-telegram-bot==12.2.0
requests==2.22.0
schedule==0.6.0
six==1.12.0
tornado==6.0.3
typed-ast==1.4.0
urllib3==1.25.6
wget==3.2
wrapt==1.11.2
zipfile38==0.0.2
